[clean] add activeChatHeader component
